Move RecentSyncLengths mocks to their own module
